+++ src/main/java/gregtech/common/gui/modularui/singleblock/base/MTETieredMachineBlockBaseGui.java
@@ -2,8 +2,15 @@ package gregtech.common.gui.modularui.singleblock.base;
 
 import static gregtech.api.metatileentity.BaseTileEntity.TOOLTIP_DELAY;
 
+import java.util.Arrays;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
+
+import org.jetbrains.annotations.NotNull;
+
 import com.cleanroommc.modularui.api.drawable.IDrawable;
 import com.cleanroommc.modularui.api.widget.IWidget;
+import com.cleanroommc.modularui.api.widget.Interactable;
 import com.cleanroommc.modularui.drawable.UITexture;
 import com.cleanroommc.modularui.factory.PosGuiData;
 import com.cleanroommc.modularui.screen.ModularPanel;
@@ -26,6 +33,7 @@ import gregtech.api.interfaces.IConfigurationCircuitSupport;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.metatileentity.implementations.MTETieredMachineBlock;
 import gregtech.api.modularui2.GTGuiTextures;
+import gregtech.api.util.GTTooltipDataCache;
 import gregtech.api.util.GTUtility;
 import gregtech.common.modularui2.factory.GTBaseGuiBuilder;
 
@@ -240,5 +248,32 @@ public class MTETieredMachineBlockBaseGui<T extends MTETieredMachineBlock> {
         tooltip.addLine(GTUtility.translate("GT5U.machines.battery_slot.tooltip"))
             .addLine(GTUtility.translate("GT5U.machines.battery_slot.tooltip.1", tierName))
             .addLine(GTUtility.translate("GT5U.machines.battery_slot.tooltip.2", tierName));
+    }
+
+    /// Sets a static tooltip using the machine's tooltip cache. This means the tooltip **can not** change while the Gui
+    /// is open.
+    ///
+    /// Accounts for extended tooltips shown when shift is held down.
+    protected Consumer<RichTooltip> configureTooltip(String key, Object... args) {
+        return addToRichTooltip(() -> machine.mTooltipCache.getData(key, args));
+    }
+
+    /// Sets a dynamic tooltip without saving it to the machine's tooltip cache. This means the tooltip **can** change
+    /// while the Gui is open.
+    ///
+    /// Accounts for extended tooltips shown when shift is held down.
+    @SafeVarargs
+    protected final Consumer<RichTooltip> configureDynamicTooltip(String key, Supplier<Object>... args) {
+        return addToRichTooltip(
+            () -> machine.mTooltipCache.getUncachedTooltipData(
+                key,
+                Arrays.stream(args)
+                    .map(Supplier::get)
+                    .toArray()));
+    }
+
+    private static @NotNull Consumer<RichTooltip> addToRichTooltip(Supplier<GTTooltipDataCache.TooltipData> data) {
+        return t -> t.addStringLines(Interactable.hasShiftDown() ? data.get().shiftText : data.get().text)
+            .titleMargin(2);
     }
 }
